+++ packages/reactabular-tree/CHANGELOG.md
@@ -1,3 +1,8 @@
+7.1.0 / 2016-xx-xx
+==================
+
+  * Bug fix - Respect `idField` properly at `tree.moveRows`.
+
 7.0.0 / 2016-11-03
 ==================